requirements.txt
discord.py
python-dotenv
a2s
flask
flask_cors